chroes: Refactor tests to use @TestExtension and avoid polluting the test Jenkins instance with test-only extensions.
