+++ src/test/java/io/jenkins/plugins/mcp/server/EndPointTest.java
@@ -29,17 +29,11 @@ package io.jenkins.plugins.mcp.server;
 import static io.jenkins.plugins.mcp.server.Endpoint.MCP_SERVER_SSE;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.jayway.jsonpath.Configuration;
-import com.jayway.jsonpath.DocumentContext;
-import com.jayway.jsonpath.JsonPath;
 import io.jenkins.plugins.mcp.server.junit.JenkinsMcpClientBuilder;
 import io.jenkins.plugins.mcp.server.junit.McpClientTest;
 import io.modelcontextprotocol.spec.McpSchema;
 import jakarta.servlet.http.HttpServletResponse;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Map;
 import org.htmlunit.HttpMethod;
 import org.htmlunit.WebRequest;
 import org.junit.jupiter.api.Test;
@@ -48,111 +42,30 @@ import org.jvnet.hudson.test.junit.jupiter.WithJenkins;
 
 @WithJenkins
 public class EndPointTest {
+    @McpClientTest
     void testListTools(JenkinsRule jenkins, JenkinsMcpClientBuilder jenkinsMcpClientBuilder) {
         try (var client = jenkinsMcpClientBuilder.jenkins(jenkins).build()) {
             client.getServerCapabilities();
             var tools = client.listTools();
-
-            var expectedFromPlugin = Arrays.asList(
-                    "whoAmI",
-                    "getBuildLog",
-                    "searchBuildLog",
-                    "triggerBuild",
-                    "updateBuild",
-                    "getJobs",
-                    "getBuild",
-                    "getJob",
-                    "getJobScm",
-                    "getBuildScm",
-                    "findJobsWithScmUrl",
-                    "getBuildChangeSets",
-                    "getStatus",
-                    "getTestResults",
-                    "getFlakyFailures",
-                    "getQueueItem");
-            var expected = new ArrayList<>();
-            expected.addAll(expectedFromPlugin);
-            expected.addAll(SampleMcpServer.getAllToolNames());
-            assertThat(tools.tools()).extracting(McpSchema.Tool::name).containsOnly(expected.toArray(new String[] {}));
-        }
-    }
-
-    @McpClientTest
-    void testSampleTool(JenkinsRule jenkins, JenkinsMcpClientBuilder jenkinsMcpClientBuilder) {
-        try (var client = jenkinsMcpClientBuilder.jenkins(jenkins).build()) {
-            var tools = client.listTools();
-            var sayHelloTool = tools.tools().stream()
-                    .filter(tool -> "sayHello".equals(tool.name()))
-                    .findFirst();
-
-            assertThat(sayHelloTool).isPresent();
-
-            assertThat(sayHelloTool.get().meta())
-                    .hasSize(2)
-                    .containsEntry("version", "1.0")
-                    .containsEntry("author", "Someone");
-
-            assertThat(sayHelloTool.get().annotations()).isNotNull();
-            assertThat(sayHelloTool.get().annotations().title()).isEqualTo("Beta tool");
-            assertThat(sayHelloTool.get().annotations().readOnlyHint()).isTrue();
-            assertThat(sayHelloTool.get().annotations().destructiveHint()).isFalse();
-            assertThat(sayHelloTool.get().annotations().idempotentHint()).isTrue();
-            assertThat(sayHelloTool.get().annotations().openWorldHint()).isFalse();
-            assertThat(sayHelloTool.get().annotations().returnDirect()).isFalse();
-
-            McpSchema.CallToolRequest request = new McpSchema.CallToolRequest("sayHello", Map.of("name", "foo"));
-
-            var response = client.callTool(request);
-            assertThat(response.isError()).isFalse();
-            assertThat(response.content()).hasSize(1);
-            assertThat(response.content().get(0).type()).isEqualTo("text");
-            assertThat(response.content()).first().isInstanceOfSatisfying(McpSchema.TextContent.class, textContent -> {
-                assertThat(textContent.type()).isEqualTo("text");
-                DocumentContext documentContext =
-                        JsonPath.using(Configuration.defaultConfiguration()).parse(textContent.text());
-
-                var contentMap = documentContext.read("$.result", Map.class);
-                assertThat(contentMap).extractingByKey("message").isEqualTo("Hello, foo!");
-            });
-        }
-    }
-
-    @McpClientTest
-    void testMcpToolCallIntResult(JenkinsRule jenkins, JenkinsMcpClientBuilder jenkinsMcpClientBuilder)
-            throws Exception {
-
-        try (var client = jenkinsMcpClientBuilder.jenkins(jenkins).build()) {
-            McpSchema.CallToolRequest request = new McpSchema.CallToolRequest("testInt", Map.of());
-
-            var response = client.callTool(request);
-            assertThat(response.isError()).isFalse();
-            assertThat(response.content()).hasSize(1);
-            assertThat(response.content().get(0).type()).isEqualTo("text");
-            assertThat(response.content()).first().isInstanceOfSatisfying(McpSchema.TextContent.class, textContent -> {
-                assertThat(textContent.type()).isEqualTo("text");
-                DocumentContext documentContext =
-                        JsonPath.using(Configuration.defaultConfiguration()).parse(textContent.text());
-                var result = documentContext.read("$.result", Integer.class);
-                assertThat(result).isEqualTo(10);
-            });
-        }
-    }
-
-    @McpClientTest
-    void testMcpToolCallWithException(JenkinsRule jenkins, JenkinsMcpClientBuilder jenkinsMcpClientBuilder)
-            throws Exception {
-
-        try (var client = jenkinsMcpClientBuilder.jenkins(jenkins).build()) {
-            McpSchema.CallToolRequest request = new McpSchema.CallToolRequest("testWithError", Map.of());
-
-            var response = client.callTool(request);
-            assertThat(response.isError()).isTrue();
-            assertThat(response.content()).hasSize(1);
-            assertThat(response.content().get(0).type()).isEqualTo("text");
-            assertThat(response.content()).first().isInstanceOfSatisfying(McpSchema.TextContent.class, textContent -> {
-                assertThat(textContent.type()).isEqualTo("text");
-                assertThat(textContent.text()).contains("Error occurred during execution");
-            });
+            assertThat(tools.tools())
+                    .extracting(McpSchema.Tool::name)
+                    .containsOnly(
+                            "whoAmI",
+                            "getBuildLog",
+                            "searchBuildLog",
+                            "triggerBuild",
+                            "updateBuild",
+                            "getJobs",
+                            "getBuild",
+                            "getJob",
+                            "getJobScm",
+                            "getBuildScm",
+                            "findJobsWithScmUrl",
+                            "getBuildChangeSets",
+                            "getStatus",
+                            "getTestResults",
+                            "getFlakyFailures",
+                            "getQueueItem");
         }
     }
